pyproject.toml
[tool.poetry]
name = "ai-value-investor"
version = "0.1.0"
description = "AI Investment Research Assistant — automated value investing research, opportunity discovery, and portfolio management"
authors = ["krantlee"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"

# ── Data Sources ──────────────────────────────────────────
akshare = "^1.14"
baostock = "^0.8"
yfinance = "^0.2"
requests = "^2.31"

# ── Data Processing ───────────────────────────────────────
pandas = "^2.2"
numpy = "^1.26"

# ── LLM Integration ───────────────────────────────────────
langchain-core = "^0.3"
langchain-openai = "^0.2"
langchain-anthropic = "^0.2"
openai = "^1.50"
anthropic = "^0.34"

# ── Data Models ───────────────────────────────────────────
pydantic = "^2.7"
pydantic-settings = "^2.5"

# ── CLI ───────────────────────────────────────────────────
click = "^8.1"
rich = "^13.7"

# ── Email ─────────────────────────────────────────────────
sib-api-v3-sdk = "^7.6"

# ── File Parsing ──────────────────────────────────────────
pdfplumber = "^0.11"

# ── Templating ────────────────────────────────────────────
jinja2 = "^3.1"

# ── Config ────────────────────────────────────────────────
python-dotenv = "^1.0"
pyyaml = "^6.0"

# ── Visualization (P2) ───────────────────────────────────
matplotlib = "^3.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
ruff = "^0.5"
mypy = "^1.10"

[tool.poetry.scripts]
invest = "src.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
